requirements.txt
# Requirements for Medical Semantic Search Engine
# Python 3.8+

# Core ML & NLP
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0

# Apache Spark (Big Data)
pyspark>=3.4.0

# Vector Search
faiss-cpu>=1.7.4
# Pour GPU: faiss-gpu>=1.7.4

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Interface
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Visualisation embeddings
umap-learn>=0.5.3

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional: Pour accélération GPU
# torch>=2.0.0
# torchvision>=0.15.0

# Development & Testing
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
